Split interop code into two separate files
